# docs/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Keycloak Multi-Tenancy",
    "version" : "23.2.0"
  },
  "servers" : [ {
    "url" : "{{keycloakUrl}}/auth/realms/{{realmName}}"
  } ],
  "paths" : {
    "/tenants" : {
      "get" : {
        "summary" : "List tenants",
        "operationId" : "listTenants",
        "parameters" : [ {
          "name" : "first",
          "in" : "query",
          "description" : "Pagination offset",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "max",
          "in" : "query",
          "description" : "Maximum results size (defaults to 100)",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "search",
          "in" : "query",
          "description" : "Tenant name",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TenantRepresentation"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          }
        }
      },
      "post" : {
        "summary" : "Create a tenant",
        "operationId" : "createTenant",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TenantRepresentation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          }
        }
      }
    },
    "/tenants/{tenantId}" : {
      "get" : {
        "summary" : "Get tenant",
        "operationId" : "getTenant",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "required" : [ "name" ],
                  "type" : "object",
                  "properties" : {
                    "id" : {
                      "type" : "string",
                      "readOnly" : true
                    },
                    "name" : {
                      "type" : "string"
                    },
                    "realm" : {
                      "type" : "string",
                      "readOnly" : true
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      },
      "put" : {
        "summary" : "Update tenant",
        "operationId" : "updateTenant",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TenantRepresentation"
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      },
      "delete" : {
        "summary" : "Delete tenant",
        "operationId" : "deleteTenant",
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          }
        }
      },
      "parameters" : [ {
        "name" : "tenantId",
        "in" : "path",
        "required" : true,
        "schema" : {
          "type" : "string"
        }
      } ]
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TenantRepresentation" : {
        "required" : [ "name" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "readOnly" : true
          },
          "name" : {
            "type" : "string"
          },
          "realm" : {
            "type" : "string",
            "readOnly" : true
          }
        }
      },
...
    }
  }
}